Copy config template to config file if it does not exist
